Port search and regex wrappers to Rust
